+++ ICD.Connect.Conferencing.Server/Devices/Client/IClientInterpretationDevice.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Conferencing.Devices;
+
+namespace ICD.Connect.Conferencing.Server.Devices.Client
+{
+	public interface IClientInterpretationDevice : IInterpretationDevice
+	{
+	}
+}