+++ ICD.Connect.Conferencing.Server/Devices/Simpl/ISimplInterpretationDeviceSettings.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Devices.Simpl;
+
+namespace ICD.Connect.Conferencing.Server.Devices.Simpl
+{
+	public interface ISimplInterpretationDeviceSettings : ISimplDeviceSettings
+	{
+	}
+}