+++ ICD.Connect.Conferencing/Controls/AbstractDialingDeviceControl.cs
@@ -17,10 +17,34 @@ namespace ICD.Connect.Conferencing.Controls
 	public abstract class AbstractDialingDeviceControl<T> : AbstractDeviceControl<T>, IDialingDeviceControl
 		where T : IDeviceBase
 	{
+		/// <summary>
+		/// Raised when a source is added to the dialing component.
+		/// </summary>
 		public abstract event EventHandler<ConferenceSourceEventArgs> OnSourceAdded;
 
+		/// <summary>
+		/// Raised when a source is removed from the dialing component.
+		/// </summary>
+		public abstract event EventHandler<ConferenceSourceEventArgs> OnSourceRemoved;
+
+		/// <summary>
+		/// Raised when a source property changes.
+		/// </summary>
+		public event EventHandler<ConferenceSourceEventArgs> OnSourceChanged;
+
+		/// <summary>
+		/// Raised when the Do Not Disturb state changes.
+		/// </summary>
 		public event EventHandler<BoolEventArgs> OnDoNotDisturbChanged;
+
+		/// <summary>
+		/// Raised when the Auto Answer state changes.
+		/// </summary>
 		public event EventHandler<BoolEventArgs> OnAutoAnswerChanged;
+
+		/// <summary>
+		/// Raised when the microphones mute state changes.
+		/// </summary>
 		public event EventHandler<BoolEventArgs> OnPrivacyMuteChanged;
 
 		private readonly SafeCriticalSection m_StateSection;
@@ -49,7 +73,7 @@ namespace ICD.Connect.Conferencing.Controls
 
 					m_AutoAnswer = value;
 
-					Logger.AddEntry(eSeverity.Informational, "{0} AutoAnswer set to {1}", this, m_AutoAnswer);
+					Log(eSeverity.Informational, "AutoAnswer set to {0}", m_AutoAnswer);
 				}
 				finally
 				{
@@ -78,7 +102,7 @@ namespace ICD.Connect.Conferencing.Controls
 
 					m_PrivacyMuted = value;
 
-					Logger.AddEntry(eSeverity.Informational, "{0} PrivacyMuted set to {1}", this, m_PrivacyMuted);
+					Log(eSeverity.Informational, "PrivacyMuted set to {0}", m_PrivacyMuted);
 				}
 				finally
 				{
@@ -107,7 +131,7 @@ namespace ICD.Connect.Conferencing.Controls
 
 					m_DoNotDisturb = value;
 
-					Logger.AddEntry(eSeverity.Informational, "{0} DoNotDisturb set to {1}", this, m_DoNotDisturb);
+					Log(eSeverity.Informational, "DoNotDisturb set to {0}", m_DoNotDisturb);
 				}
 				finally
 				{
@@ -187,6 +211,41 @@ namespace ICD.Connect.Conferencing.Controls
 		/// </summary>
 		/// <param name="enabled"></param>
 		public abstract void SetPrivacyMute(bool enabled);
+
+		#endregion
+
+		#region Source Events
+
+		/// <summary>
+		/// Subscribes to the source events in order to re-raise OnSourceChanged event.
+		/// </summary>
+		/// <param name="source"></param>
+		protected void SourceSubscribe(IConferenceSource source)
+		{
+			source.OnAnswerStateChanged += SourceOnPropertyChanged;
+			source.OnNameChanged += SourceOnPropertyChanged;
+			source.OnNumberChanged += SourceOnPropertyChanged;
+			source.OnSourceTypeChanged += SourceOnPropertyChanged;
+			source.OnStatusChanged += SourceOnPropertyChanged;
+		}
+
+		/// <summary>
+		/// Unsubscribes from the source events to stop re-raising OnSourceChanged event.
+		/// </summary>
+		/// <param name="source"></param>
+		protected void SourceUnsubscribe(IConferenceSource source)
+		{
+			source.OnAnswerStateChanged -= SourceOnPropertyChanged;
+			source.OnNameChanged -= SourceOnPropertyChanged;
+			source.OnNumberChanged -= SourceOnPropertyChanged;
+			source.OnSourceTypeChanged -= SourceOnPropertyChanged;
+			source.OnStatusChanged -= SourceOnPropertyChanged;
+		}
+
+		private void SourceOnPropertyChanged(object sender, EventArgs args)
+		{
+			OnSourceChanged.Raise(this, new ConferenceSourceEventArgs(sender as IConferenceSource));
+		}
 
 		#endregion