+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [7.0.0] - 2018-05-24
+### Changed
+ - Significant Interpretation device refactoring
+ 
+## [6.0.0] - 2018-05-09
+### Added
+ - Adding Interpretation client and server devices
+ - Adding Interpretation shims for S+
+
+### Changed
+ - Moved codec input types configuration into new CodecInputTypes class
+
+## [5.2.0] - 2018-05-03
 ### Added
  - Additional console commands for dialing controls